requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pydantic[email]==2.11.4
pydantic-settings==2.9.1
python-multipart==0.0.20
sqlmodel==0.0.24
bcrypt==4.3.0
python-jose==3.4.0
psycopg2-binary==2.9.10
requests==2.32.3
unique-namer==1.6.1
pytest==8.3.5
httpx==0.28.1
